refactor: centralize repository URL classification
